first draft with shitty UI layout
