+++ src/components/Flashcard/types.ts
@@ -1,0 +1,9 @@
+export interface FlashcardQAPairUI {
+  question: string;
+  answer: string;
+  isFlipped: boolean;
+}
+export interface FlashcardQAPair {
+  question: string;
+  answer: string;
+}
